requirements.txt
streamlit
langchain
langgraph
langchain-google-genai
google-generativeai
python-dotenv
pytesseract
pymupdf
pandas
openpyxl
Pillow
